stuNumber change from int to long not test yet

+++ src/main/java/com/communitystreet/dao/CommunityDao.java
@@ -13,12 +13,12 @@ import java.util.List;
 @Repository
 public interface CommunityDao {
 //    Community getCommunityById(@Param("stuNumber") int stuNumber);
-    Community getCommunityByAdmin(@Param("stuNumber") int stuNumber);
+    Community getCommunityByAdmin(@Param("stuNumber") long stuNumber);
     List<Community> getCommnuityByName(@Param("communityName") String communityName);
-    int createCommunity(Community community, @Param("stuNumber") int stuNumber);
+    int createCommunity(Community community, @Param("stuNumber") long stuNumber);
     int userJoinCommunity(@Param("user") User user, @Param("community") Community community);
     // /社团界面的显示社团
     List<Community> getAll();
     // /个人界面中的显示个人加入的社团
-    List<Community> getCommunities(@Param("stuNumber") int stuNumber);
+    List<Community> getCommunities(@Param("stuNumber") long stuNumber);
 }
